if no products query string then load the products only once and reuse it
