+++ misc/services/mobile-webui/mobile-webui-frontend/src/apps/pos/reducers/posTerminal.js
@@ -84,7 +84,6 @@ const setPOSTerminalToState = ({ applicationState, newTerminal }) => {
     isLoaded: true,
   };
   delete newTerminalEffective.openOrders;
-  delete newTerminalEffective.products;
 
   return {
     ...applicationState,
